Fix segfaults inside g_free on windows
calling `libc::free` actually doesn't always work, because there are
multiple heaps, and we're trying to free it from the wrong one.

Instead, we find the g_free export from libglib-2.0, and invoke that.

Added the `libloading` crate to help with this, and also rewrote the
`delaylink_hook` to use the functions there, rather than directly
calling the windows API.